fix: test 오류 개선
