+++ packages/core/src/tool/mcp/registry/__tests__/mcp-metadata-registry.test.ts
@@ -1,151 +1,24 @@
-import { McpMetadataRegistry } from '../mcp-metadata-registry';
-import {
-  McpToolRepository,
-  McpToolRepositoryEventPayload,
-} from '../../repository/mcp-tool-repository';
-import { McpToolMetadata, McpConnectionStatus } from '../../mcp-types';
+import { mock } from 'jest-mock-extended';
+import { McpRegistry } from '../..';
 import { McpConfig } from '../../mcp-config';
-import {
-  CursorPagination,
-  CursorPaginationResult,
-} from '../../../../common/pagination/cursor-pagination';
-
-// Mock Repository 구현
-class MockMcpToolRepository implements McpToolRepository {
-  private tools = new Map<string, McpToolMetadata>();
-  private eventHandlers = new Map<string, ((payload: McpToolRepositoryEventPayload) => void)[]>();
-
-  async get(id: string): Promise<McpToolMetadata | null> {
-    return this.tools.get(id) || null;
-  }
-
-  async list(pagination?: CursorPagination): Promise<CursorPaginationResult<McpToolMetadata>> {
-    const items = Array.from(this.tools.values());
-    return {
-      items,
-      nextCursor: '',
-      hasMore: false,
-    };
-  }
-
-  async search(): Promise<CursorPaginationResult<McpToolMetadata>> {
-    return this.list();
-  }
-
-  async create(config: McpConfig): Promise<McpToolMetadata> {
-    const tool: McpToolMetadata = {
-      id: `tool_${Date.now()}_${Math.random().toString(36).substr(2, 9)}`,
-      name: config.name,
-      description: `MCP Tool: ${config.name}`,
-      version: config.version,
-      category: 'general',
-      provider: 'Test Provider',
-      status: 'disconnected',
-      usageCount: 0,
-      permissions: [],
-      config: {},
-    };
-
-    this.tools.set(tool.id, tool);
-    this.emit('changed', { id: tool.id, metadata: tool });
-
-    return tool;
-  }
-
-  async update(
-    id: string,
-    patch: Partial<McpToolMetadata>,
-    options?: { expectedVersion?: string }
-  ): Promise<McpToolMetadata> {
-    const existing = this.tools.get(id);
-    if (!existing) {
-      throw new Error(`Tool not found: ${id}`);
-    }
-
-    if (options?.expectedVersion && existing.version !== options.expectedVersion) {
-      throw new Error(`Version conflict for tool ${id}`);
-    }
-
-    const previousStatus = existing.status;
-    const updated: McpToolMetadata = {
-      ...existing,
-      ...patch,
-      id, // ID는 변경 불가
-      version: `v${Date.now()}`, // 새 버전 할당
-    };
-
-    this.tools.set(id, updated);
-    this.emit('changed', { id, metadata: updated });
-
-    if (previousStatus !== updated.status) {
-      this.emit('statusChanged', { id, metadata: updated, previousStatus });
-    }
-
-    return updated;
-  }
-
-  async delete(id: string): Promise<void> {
-    if (!this.tools.has(id)) {
-      throw new Error(`Tool not found: ${id}`);
-    }
-
-    this.tools.delete(id);
-    this.emit('deleted', { id });
-  }
-
-  on(
-    event: 'changed' | 'deleted' | 'statusChanged',
-    handler: (payload: McpToolRepositoryEventPayload) => void
-  ): () => void {
-    if (!this.eventHandlers.has(event)) {
-      this.eventHandlers.set(event, []);
-    }
-
-    this.eventHandlers.get(event)!.push(handler);
-
-    return () => {
-      const handlers = this.eventHandlers.get(event)!;
-      const index = handlers.indexOf(handler);
-      if (index >= 0) {
-        handlers.splice(index, 1);
-      }
-    };
-  }
-
-  private emit(event: string, payload: McpToolRepositoryEventPayload): void {
-    const handlers = this.eventHandlers.get(event) || [];
-    handlers.forEach((handler) => handler(payload));
-  }
-}
-
-// MCP 연결 실패를 시뮬레이션하기 위한 Mock
-jest.mock('../mcp-metadata-registry', () => {
-  const actual = jest.requireActual('../mcp-metadata-registry');
-  return {
-    ...actual,
-    McpMetadataRegistry: class extends actual.McpMetadataRegistry {
-      // MCP Registry mock을 위해 생성자에서 실제 연결을 하지 않도록 override
-      constructor(repository: any) {
-        super(repository);
-        // 실제 MCP 연결을 모킹하기 위해 mcpRegistry를 mock으로 교체할 수 있음
-      }
-    },
-  };
-});
+import { McpMetadataRegistry } from '../mcp-metadata-registry';
+import { MockMcpToolRepository } from './fixture';
 
 describe('McpMetadataRegistry', () => {
   let registry: McpMetadataRegistry;
   let mockRepository: MockMcpToolRepository;
+  let mockMcpRegistry: ReturnType<typeof mock<McpRegistry>>;
 
   beforeEach(async () => {
+    mockMcpRegistry = mock<McpRegistry>();
     mockRepository = new MockMcpToolRepository();
-    registry = new McpMetadataRegistry(mockRepository);
+    registry = new McpMetadataRegistry(mockRepository, mockMcpRegistry);
     await registry.initialize();
   });
 
   describe('initialization', () => {
     it('should initialize successfully', async () => {
-      const newRegistry = new McpMetadataRegistry(mockRepository);
+      const newRegistry = new McpMetadataRegistry(mockRepository, mockMcpRegistry);
       await expect(newRegistry.initialize()).resolves.toBeUndefined();
     });
 
@@ -160,7 +33,7 @@ describe('McpMetadataRegistry', () => {
       await mockRepository.create(config);
 
       // 새 레지스트리 인스턴스로 초기화
-      const newRegistry = new McpMetadataRegistry(mockRepository);
+      const newRegistry = new McpMetadataRegistry(mockRepository, mockMcpRegistry);
       await newRegistry.initialize();
 
       expect(newRegistry.totalToolsCount).toBe(1);
@@ -175,6 +48,8 @@ describe('McpMetadataRegistry', () => {
         version: '1.0.0',
         command: 'node',
       };
+
+      mockMcpRegistry.register.mockRejectedValue(new Error('Connection failed'));
 
       // MCP 연결 실패를 시뮬레이션 (실제 MCP 서버가 없으므로)
       await expect(registry.registerTool(config)).rejects.toThrow();
@@ -352,12 +227,12 @@ describe('McpMetadataRegistry', () => {
 
     it('should filter tools by status', () => {
       const errorTools = registry.getToolsByStatus('error');
-      expect(errorTools).toHaveLength(3); // 모든 도구가 연결 실패로 error 상태
+      expect(errorTools).toHaveLength(0);
     });
 
     it('should provide counts', () => {
       expect(registry.totalToolsCount).toBe(3);
-      expect(registry.connectedToolsCount).toBe(0); // 실제 연결된 도구 없음
+      expect(registry.connectedToolsCount).toBe(3);
     });
   });
