+++ packages/core/tsconfig.json
@@ -4,5 +4,5 @@
     "rootDir": "./src"
   },
   "include": ["src"],
-  "exclude": ["node_modules", "dist", "esm", "src/**/__tests__"]
+  "exclude": ["node_modules", "dist", "esm"]
 }
